requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==17.6.0
inflection==0.5.1
iniconfig==2.0.0
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
psycopg2==2.9.5
PyJWT==2.6.0
pytest==7.2.2
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
